pyproject.toml
[tool.poetry]
name = "legoukrainebot"
version = "0.0.0"
description = "Lego Ukraine Discord bot"
authors = ["Hhummann <hhummannv@gmail.com>"]
license = "Emerald Wave"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
